Move ScriptException from .core to .script package.
