+++ core/src/main/java/org/bitcoinj/script/ScriptException.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package org.bitcoinj.core;
+package org.bitcoinj.script;
 
-import org.bitcoinj.script.ScriptError;
+import org.bitcoinj.core.VerificationException;
 
 @SuppressWarnings("serial")
 public class ScriptException extends VerificationException {
